RM#58656 Remove changing product name on an invoice line in qty onchange (#9563)
(cherry picked from commit bf8cc4dba812e4056f0e247b93540d26525e8ea3)

+++ axelor-purchase/src/main/java/com/axelor/apps/purchase/service/SupplierCatalogServiceImpl.java
@@ -238,7 +238,8 @@ public class SupplierCatalogServiceImpl implements SupplierCatalogService {
 
     if (qty.compareTo(minQty) < 0) {
       String msg =
-          String.format(I18n.get(PurchaseExceptionMessage.PURCHASE_ORDER_LINE_MIN_QTY), minQty);
+          String.format(
+              I18n.get(PurchaseExceptionMessage.PURCHASE_ORDER_LINE_MIN_QTY), minQty.intValue());
 
       if (request.getAction().endsWith("onchange")) {
         response.setFlash(msg);
